refactor: move predictor functions to prediction.py
